requirements.txt
python-dotenv
pygame-ce
numpy